% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read_write.R
\name{dataf_read}
\alias{dataf_read}
\title{Data frame read and write}
\usage{
dataf_read(), dataf_write()
}
\arguments{
\item{one}{dataframe or a path where data frame stored}
}
\value{
a data frame read from files
}
\description{
Data frame read and write
}
