% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_detect.R
\name{fac_detect}
\alias{fac_detect}
\title{Factors detect whether the input is a character}
\usage{
fac_detect()
}
\arguments{
\item{one}{factor}
}
\value{
TRUE(factor), FALSE(character)
}
\description{
If the input is a character, output FALSE
If the input is a factor, output TRUE
}
