% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_reorder.R
\name{fac_reorder}
\alias{fac_reorder}
\title{Reorder factors in a descending order}
\usage{
fac_reorder()
}
\arguments{
\item{one}{factor}
}
\value{
same factor but in descending order
}
\description{
Reorder factors in a descending order
}
