% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_set.R
\name{fac_set}
\alias{fac_set}
\title{Setting levels as it appears in data}
\usage{
fac_set()
}
\arguments{
\item{one}{factor}
}
\value{
same factor but whose level set to the order of itself
}
\description{
Setting levels as it appears in data
}
